requirements/development.txt
Django>=3.2,<5.0
Pillow>=10.1.0,<10.2.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.3.1,<5.4.0
psycopg2-binary>=2.9.9,<2.10.0
